python/requirements.txt
msgraph-sdk
requests
azure-identity
